Package renamed + firebase setup.
