+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.appcontrolecalorico",
+    package = "com.santos.fellyp.contadorcalorias",
 )
 
 android_resource(
     name = "res",
-    package = "com.appcontrolecalorico",
+    package = "com.santos.fellyp.contadorcalorias",
     res = "src/main/res",
 )
