+++ android/app/src/main/java/com/santos/fellyp/contadorcalorias/MainActivity.java
@@ -1,4 +1,4 @@
-package com.appcontrolecalorico;
+package com.santos.fellyp.contadorcalorias;
 
 import com.facebook.react.ReactActivity;
