+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">appcontrolecalorico</string>
+    <string name="app_name">Contador de Calorias</string>
 </resources>
