+++ android/build.gradle
@@ -14,6 +14,7 @@ buildscript {
     }
     dependencies {
         classpath("com.android.tools.build:gradle:4.1.0")
+        classpath("com.google.gms:google-services:4.3.8")
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
