+++ app.json
@@ -1,4 +1,4 @@
 {
   "name": "appcontrolecalorico",
-  "displayName": "appcontrolecalorico"
+  "displayName": "Contador de Calorias"
 }
