+++ package.json
@@ -10,6 +10,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-firebase/app": "^12.3.0",
     "react": "17.0.1",
     "react-native": "0.64.2"
   },
